requirements.txt
click==6.6
dask==0.10.1
decorator==4.0.10
dlib==18.17.100
Flask==0.11.1
Flask-SocketIO==2.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
networkx==1.11
numpy==1.11.1
Pillow==3.3.0
python-engineio==0.9.2
python-socketio==1.4.2
scikit-image==0.12.3
scipy==0.17.1
six==1.10.0
toolz==0.8.0
Werkzeug==0.11.10
